+++ src/features/favorite-toggle/index.ts
@@ -1,0 +1,2 @@
+export { FavoriteButton } from '#/features/favorite-toggle/ui/FavoriteButton'
+export { useToggleFavorite, useIsFavorite } from '#/features/favorite-toggle/model/use-favorite'